+++ ressources/webpack-entry-points/browser-script.cjs
@@ -1,0 +1,4 @@
+/* eslint-disable no-undef */
+"use strict";
+
+window[process.env.__PACKAGE_NAME__] = require(process.env.__PACKAGE_ROOT__).default;