+++ src/Manager.js
@@ -1,7 +1,7 @@
 //--------------------------------------------------------
 //-- Manager
 //--------------------------------------------------------
-import brand from "@absolunet/brand-guidelines";
+import { guidelines } from "@absolunet/brand-guidelines";
 import { Joi, validateArgument } from "@absolunet/joi";
 import { terminal } from "@absolunet/terminal";
 import emoji from "node-emoji";
@@ -20,8 +20,8 @@ class Manager {
 	 * Create a Manager.
 	 */
 	constructor() {
-		const mainColor = brand.styleguide.color.greyscale.nevada;
-		const secondaryColor = brand.styleguide.color.greyscale.geyser;
+		const mainColor = guidelines.color.achromatic.gray;
+		const secondaryColor = guidelines.color.achromatic.gray;
 
 		terminal.setTheme({
 			logo: emoji.get("technologist"), // 🧑‍💻
@@ -31,20 +31,6 @@ class Manager {
 			borderColor: mainColor,
 			spinnerColor: terminal.basicColor.grey,
 		});
-	}
-
-	/**
-	 * Update package meta.
-	 *
-	 * @async
-	 * @param {string} [absolutePath={@link PackagePaths}.root] - Directory path of license.
-	 * @returns {Promise} When method completed.
-	 */
-	// eslint-disable-next-line require-await
-	async updatePackageMeta(absolutePath) {
-		validateArgument("absolutePath", absolutePath, Joi.absolutePath());
-
-		util.updateLicense(absolutePath);
 	}
 
 	/**
@@ -61,19 +47,6 @@ class Manager {
 	}
 
 	/**
-	 * Reinstall packages.
-	 *
-	 * @async
-	 * @param {string} [absolutePath={@link PackagePaths}.root] - Directory path of the package.json.
-	 * @returns {Promise} When method completed.
-	 */
-	async installPackage(absolutePath) {
-		validateArgument("absolutePath", absolutePath, Joi.absolutePath());
-
-		await util.npmInstall(absolutePath);
-	}
-
-	/**
 	 * Initialize the manager.
 	 *
 	 * @async
@@ -81,19 +54,22 @@ class Manager {
 	 * @returns {Promise} When method completed.
 	 * @example
 	 * manager.init({
-	 * 	repositoryType: 'single-package',
+	 * 	repositoryType: "single-package",
+	 * 	dist: {
+	 *  	node: {},
+	 *  },
 	 * 	tasks: {
 	 *		build: {
-	 *			postRun: async () => {}
+	 *			postRun: async () => {},
 	 * 		},
 	 *
-	 * 		deploy: {
+	 * 		prepare: {
 	 * 			preRun:  async () => {},
 	 * 			postRun: async ({ terminal }) => {
-	 * 				terminal.print('Enjoy');
-	 * 			}
-	 * 		}
-	 * 	}
+	 * 				terminal.print("Enjoy");
+	 * 			},
+	 * 		},
+	 * 	},
 	 * });
 	 */
 	async init(options = {}) {
@@ -102,27 +78,36 @@ class Manager {
 			options,
 			Joi.object({
 				repositoryType: Joi.string().valid(...Object.values(environment.REPOSITORY_TYPE)),
-				restricted: Joi.boolean(),
-				useOTP: Joi.boolean(),
 
 				dist: Joi.object({
 					source: Joi.absolutePath(),
 					destination: Joi.absolutePath(),
-					node: Joi.boolean(),
-					web: Joi.object({
-						types: Joi.array()
-							.items(Joi.string().valid(...Object.values(environment.DISTRIBUTION_WEB_TYPE)))
-							.min(1)
-							.unique()
-							.required(),
-						name: Joi.variableName().required(),
-						externals: Joi.object().pattern(/^[a-z0-9-/@]$/iu, Joi.variableName()),
+					node: Joi.object({
+						type: Joi.string().valid(...Object.values(environment.DISTRIBUTION_NODE_TYPE)),
+						target: Joi.string().empty(),
 					}),
+					browser: Joi.array()
+						.items(
+							Joi.object({
+								type: Joi.string()
+									.valid(...Object.values(environment.DISTRIBUTION_BROWSER_TYPE))
+									.required(),
+								target: Joi.string().empty(),
+								name: Joi.variableName().when("type", {
+									is: Joi.valid(environment.DISTRIBUTION_BROWSER_TYPE.script),
+									then: Joi.required(),
+								}),
+								externals: Joi.object().pattern(/^[a-z0-9/@-]+$/iu, Joi.variableName()),
+							})
+						)
+						.min(1)
+						.unique(),
 					include: Joi.array().items(Joi.string()),
-				}).required(),
+				})
+					.or("node", "browser")
+					.required(),
 
 				tasks: Joi.object(
-					// eslint-disable-next-line unicorn/prefer-object-from-entries
 					Object.values(environment.TASK).reduce((list, task) => {
 						list[task] = {
 							preRun: Joi.function(),
@@ -147,9 +132,6 @@ class Manager {
 		}
 
 		switch (util.getTask()) {
-			case environment.TASK.install:
-				await managerType.install();
-				break;
 			case environment.TASK.outdated:
 				await managerType.outdated();
 				break;
@@ -169,17 +151,11 @@ class Manager {
 			case environment.TASK.prepare:
 				await managerType.prepare();
 				break;
+			case environment.TASK.version:
+				await managerType.version();
+				break;
 			case environment.TASK.rebuild:
 				await managerType.rebuild();
-				break;
-
-			case environment.TASK.publish:
-				await managerType.publish();
-				break;
-
-			case environment.TASK.publishUnsafe:
-				await util.confirmUnsafePublish();
-				await managerType.publish({ unsafe: true });
 				break;
 
 			default: